Use Long for GitHub generated bookkeeping ids

+++ src/main/java/com/github/shredder121/gh_event_api/handler/page_build/PageBuildPayload.java
@@ -42,7 +42,7 @@ public class PageBuildPayload {
     /**
      * The id of the page build, for GitHub bookkeeping.
      */
-    @NotNull Integer id;
+    @NotNull Long id;
 
     /**
      * The page build itself.
